+++ development/circular-deps.jsonc
@@ -51,12 +51,5 @@
   [
     "ui/pages/notifications/notifications-list.tsx",
     "ui/pages/notifications/notifications.tsx"
-  ],
-  [
-    "ui/pages/snap-account-redirect/components/index.ts",
-    "ui/pages/snap-account-redirect/components/snap-account-redirect-context.tsx",
-    "ui/pages/snap-account-redirect/components/snap-account-redirect-message.tsx",
-    "ui/pages/snap-account-redirect/components/url-display-box.tsx",
-    "ui/pages/snap-account-redirect/snap-account-redirect.tsx"
   ]
 ]
