+++ ui/pages/snap-account-redirect/snap-account-redirect.tsx
@@ -8,7 +8,7 @@ import {
   FlexDirection,
 } from '../../helpers/constants/design-system';
 import SnapAuthorshipHeader from '../../components/app/snaps/snap-authorship-header';
-import { SnapAccountRedirectContent } from './components';
+import SnapAccountRedirectContent from './components/snap-account-redirect-context';
 
 export type SnapAccountRedirectProps = {
   url: string;
